n2_mobile/requirements.txt
Django
psycopg2-binary
django-redis
gunicorn
djangorestframework